Changes json in eirebase and added filter
